V3.78: adding a double-double arithmetic type (#426)
* Setting SEMVER to v3.78

* Setting SEMVER to v3.78

* setting default devcontainer to clang16

* header updates for elastic integer

* header update for c_api

* header update for elastic float

* header update for elastic posit

* header updates for elastic unum

* header updates of mixed-precision applications

* header updates

* adding mathematics section to accuracy applications

* std::numbers inspired numerical constants, sans the C++20 concepts

* AppleClang compilation fix

* header changes for applications of accuracy

* header update and code hygiene for precision applications

* header updates for applications generating mathematical constants

* creating a mixed-precision midpoint/lerp skeleton test to lerp test complex ranges

* code hygiene on precision applications

* header updates and code hygiene for precision applications

* header updates and code hygiene for rest of application projects

* residual compensation algorithm using the quire

* code hygiene for dfloat skeleton

* code hygiene for dfloat skeleton

* code hygiene for bfloat16 skeleton

* initial skeleton for doubledouble arithmetic type

* gcc/clang compilation fix for dfloat template

* adding a doubledouble skeleton

* adding missing quad-double directory and build file

* adding maxpos calculation to doubledouble

* WIP: bootstrapping doubledouble algorithms

* WIP: experiments how to verify doubledouble arithmetic correctness

* compilation and test fixes of doubledouble regression tests

* disabling doubledouble regression tests until coverage

* WIP: bug fix in quad-double multiply operator

* WIP: implementing numerics and classification functions for doubledouble

* WIP: decimal string printing and parsing of doubledouble

* replacing MSVC specific _fpclass API with standard C++ for doubledouble fpclassify

* WIP: skeletons for doubledouble math function regression tests

* bug fix for compilation guard logic for checkNaN/checkInf constexpr functions

* removing BIT_CAST_SUPPORT and standardize on bit_cast.hpp auto configuration

* removing constexpr from bfloat[8|16].assign() as clang doesn't have constexpr std::string

* second attempt to get the BIT_CAST_CONSTEXPR compiler guards correct across the three compilers

* third attempt to get BIT_CAST and LONG_DOUBLE support working across all compilers

* defeaturing std::bit_cast<> from processing long doubles

* removing std::bit_cast<blob, long double>

* adding doubledouble constant generation example

* streamlining doubledouble api tests

* adding a skeleton for experimentation with doubledouble

* changing to_binary of a doubledouble to reflect the hidden bit of the second segment

* adding performance measurements for dd vs native and quad emulation

* marking all variables and intermediates in the Kahan sums and products volatile

* compilation fix for gcc

* experiment with infinite sequence constants

* updating build containers to contain gdb

* code hygiene on volatile qualifier

* implementing setbit on double-double

* WIP: std::stream functionality for double-double

* new utility to report on language features configured

* removing <format> as gcc doesn't support it yet

* quick readme update

* making <stdfloat> conditional as gcc and clang do not support it yet

* code hygiene for double-double

* completing arithmetic regression test skeletons for double-double

* removing unsafe sscanf from parse function of double-double

* WIP: code streamlining and documentation

* code hygiene

* adding pow, exp, and sqrt algorithms for double-double courtesy of Scibuilders

* adding double-double constants of note and a generator

* generating exp table

* code hygiene

* adding shims for the rest of the math library for double-double

* compilation fix for gcc and clang

+++ include/universal/blas/solvers/lu.hpp
@@ -87,10 +87,10 @@ void Crout(const Matrix& S, Matrix& D) {
 template<typename Matrix, typename Vector>
 void SolveCrout(const Matrix& LU, const Vector& b, Vector& x) {
 	assert(num_cols(LU) == size(b));
-	size_t N = size(b);
+	unsigned N = size(b);
 	using value_type = typename Matrix::value_type;
 	sw::universal::blas::vector<value_type> y(N);
-	for (size_t i = 0; i < N; ++i) {
+	for (unsigned i = 0; i < N; ++i) {
 		value_type sum = 0.0;
 		for (size_t k = 0; k < size_t(i); ++k) sum += LU[i][k] * y[k];
 		y[i] = (b[i] - sum) / LU[i][i];
@@ -98,7 +98,7 @@ void SolveCrout(const Matrix& LU, const Vector& b, Vector& x) {
 	}
 	for (int i = static_cast<int>(N) - 1; i >= 0; --i) {
 		value_type sum = 0.0;
-		for (int k = i + 1; k < static_cast<int>(N); ++k) {
+		for (unsigned k = i + 1; k < N; ++k) {
 			//cout << "lu[] = " << LU[i][k] << " x[" << k << "] = " << x[k] << endl;
 			sum += LU[i][k] * x[k];
 		}
